backend/requirements.txt
fastapi
uvicorn
python-docx
google-generativeai
python-multipart
Pillow
pydantic
python-dotenv
